Broke out exceptions.h from worker.h
