+++ src/odin/maneuversbuilder.cc
@@ -6,6 +6,7 @@
 #include "baldr/turn.h"
 #include "baldr/turnlanes.h"
 #include "baldr/verbal_text_formatter_factory.h"
+#include "exceptions.h"
 #include "midgard/logging.h"
 #include "midgard/util.h"
 #include "odin/sign.h"
@@ -13,7 +14,6 @@
 #include "odin/util.h"
 #include "proto/directions.pb.h"
 #include "proto/options.pb.h"
-#include "worker.h"
 
 #ifdef LOGGING_LEVEL_DEBUG
 #include "midgard/encoded.h"
