+++ src/sif/transitcost.cc
@@ -2,10 +2,10 @@
 #include "baldr/graphconstants.h"
 #include "baldr/rapidjson_utils.h"
 #include "proto_conversions.h"
-#include "worker.h"
 
 #ifdef INLINE_TEST
 #include "test.h"
+#include "worker.h"
 
 #include <random>
 #endif
