+++ src/thor/costmatrix.cc
@@ -1,10 +1,10 @@
 #include "thor/costmatrix.h"
 #include "baldr/datetime.h"
+#include "exceptions.h"
 #include "midgard/encoded.h"
 #include "midgard/logging.h"
 #include "sif/hierarchylimits.h"
 #include "sif/recost.h"
-#include "worker.h"
 
 #include <ankerl/unordered_dense.h>
