+++ test/gurka/test_access.cc
@@ -1,5 +1,5 @@
+#include "exceptions.h"
 #include "gurka.h"
-#include "worker.h"
 
 #include <gtest/gtest.h>
