+++ valhalla/odin/worker.h
@@ -1,6 +1,7 @@
 #ifndef __VALHALLA_ODIN_SERVICE_H__
 #define __VALHALLA_ODIN_SERVICE_H__
 
+#include <valhalla/exceptions.h>
 #include <valhalla/odin/markup_formatter.h>
 #include <valhalla/proto/api.pb.h>
 #include <valhalla/worker.h>
