+++ valhalla/thor/worker.h
@@ -3,6 +3,7 @@
 
 #include <valhalla/baldr/attributes_controller.h>
 #include <valhalla/baldr/graphreader.h>
+#include <valhalla/exceptions.h>
 #include <valhalla/meili/map_matcher_factory.h>
 #include <valhalla/meili/match_result.h>
 #include <valhalla/proto/options.pb.h>
