+++ valhalla/worker.h
@@ -1,5 +1,6 @@
-#ifndef __VALHALLA_SERVICE_H__
-#define __VALHALLA_SERVICE_H__
+#ifndef __VALHALLA_WORKER_H__
+#define __VALHALLA_WORKER_H__
+#include <valhalla/exceptions.h>
 #include <valhalla/proto/api.pb.h>
 #include <valhalla/sif/dynamiccost.h>
 
@@ -17,37 +18,6 @@ namespace valhalla {
 struct hierarchy_limits_config_t {
   std::vector<HierarchyLimits> max_limits;
   std::vector<HierarchyLimits> default_limits;
-};
-
-/**
- * Project specific error messages and codes that can be converted to http responses
- */
-struct valhalla_exception_t : public std::runtime_error {
-  /**
-   * Constructs the exception by looking up predefined ones by their codes. If unsuccessful the code
-   * will be 0
-   * @param code   the code to look up
-   * @param extra  an extra string to append to the codes existing method
-   */
-  valhalla_exception_t(unsigned code, const std::string& extra = "");
-  valhalla_exception_t(unsigned code,
-                       const std::string& message,
-                       unsigned http_code,
-                       const std::string& http_message,
-                       const std::string& osrm_error,
-                       const std::string& statsd_key = "")
-      : std::runtime_error(""), code(code), message(message), http_code(http_code),
-        http_message(http_message), osrm_error(osrm_error), statsd_key(statsd_key) {
-  }
-  const char* what() const noexcept override {
-    return message.c_str();
-  }
-  unsigned code;
-  std::string message;
-  unsigned http_code;
-  std::string http_message;
-  std::string osrm_error;
-  std::string statsd_key;
 };
 
 /**
@@ -203,4 +173,4 @@ protected:
 };
 } // namespace valhalla
 
-#endif //__VALHALLA_SERVICE_H__
+#endif //__VALHALLA_WORKER_H__
